deployment/requirements.txt
streamlit == 1.28.2
pandas == 2.1.1
seaborn == 0.13.0
matplotlib == 3.8.1
numpy == 1.26.1
scikit-learn == 1.3.2
plotly == 5.18.0
Pillow == 10.0.1
imblearn == 0.0